move utilites. make train simple to work with
